refactor: extract AudienceDetailTabs into separate component
